EZP-24379: Abstracted action dispatcher for better reusability
